+++ PSFramework/functions/utility/Join-PSFPath.ps1
@@ -35,6 +35,7 @@
 		$Path,
 		
 		[Parameter(ValueFromRemainingArguments = $true)]
+		[Alias('ChildPath')]
 		[string[]]
 		$Child,