+++ library/PSFramework/Runspace/RunspaceBoundValueGeneric.cs
@@ -50,7 +50,7 @@ namespace PSFramework.Runspace
         /// <summary>
         /// Removes all value entries whose corresponding Runspace has been destroyed
         /// </summary>
-        public new void PurgeExpired()
+        public override void PurgeExpired()
         {
             // Store IDs first, so parallel access is not an issue and a new value gets accidentally discarded
             Guid[] IDs = Values.Keys.ToList().ToArray();